# docs/swagger.yaml
basePath: /api
definitions:
  entity.PetType:
    enum:
    - dog
    - cat
    - bird
    - fish
    - rabbit
    type: string
    x-enum-varnames:
    - Dog
    - Cat
    - Bird
    - Fish
    - Rabbit
  helper.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  v1.CreateOwnerBody:
    properties:
      name:
        type: string
    type: object
  v1.CreateOwnerResponse:
    properties:
      message:
        type: string
    type: object
  v1.CreatePetBody:
    properties:
      name:
        type: string
      owner_id:
        type: string
      type:
        type: string
    type: object
  v1.CreatePetResponse:
    properties:
      message:
        type: string
    type: object
  v1.DeleteOwnerResponse:
    properties:
      message:
        type: string
    type: object
  v1.DeletePetResponse:
    properties:
      message:
        type: string
    type: object
  v1.GetOwnerResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      updated_at:
        type: integer
    type: object
  v1.GetPetResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      type:
        $ref: '#/definitions/entity.PetType'
      updated_at:
        type: integer
    type: object
  v1.ListOwnersResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      rows:
        items:
          $ref: '#/definitions/v1.GetOwnerResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  v1.ListPetsResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      rows:
        items:
          $ref: '#/definitions/v1.GetPetResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  v1.UpdateOwnerBody:
    properties:
      name:
        type: string
    type: object
  v1.UpdateOwnerResponse:
    properties:
      message:
        type: string
    type: object
  v1.UpdatePetBody:
    properties:
      name:
        type: string
      owner_id:
        type: string
      type:
        type: string
    type: object
  v1.UpdatePetResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a boilerplate for microservice using golang, grpc, and http.
  title: go-microservice-boilerplate API Documentation
  version: "1.0"
paths:
  /v1/owners:
    get:
      consumes:
      - application/json
      description: List owners
      operationId: ListOwners
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per Page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListOwnersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: List owners
      tags:
      - owners v1
    post:
      consumes:
      - application/json
      description: Create a owner
      operationId: CreateOwner
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateOwnerBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CreateOwnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Create a owner
      tags:
      - owners v1
  /v1/owners/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a owner
      operationId: DeleteOwner
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.DeleteOwnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Delete a owner
      tags:
      - owners v1
    get:
      consumes:
      - application/json
      description: Get a Owner
      operationId: GetOwner
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetOwnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get a Owner
      tags:
      - owners v1
    patch:
      consumes:
      - application/json
      description: Update a owner
      operationId: UpdateOwner
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateOwnerBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateOwnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Update a owner
      tags:
      - owners v1
  /v1/pets:
    get:
      consumes:
      - application/json
      description: List Pets
      operationId: ListPets
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per Page
        in: query
        name: per_page
        type: integer
      - description: Has Owner
        in: query
        name: has_owner
        type: boolean
      - description: Owner ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListPetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: List Pets
      tags:
      - pets v1
    post:
      consumes:
      - application/json
      description: Create a pet
      operationId: CreatePet
      parameters:
      - description: Name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreatePetBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CreatePetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Create a pet
      tags:
      - pets v1
  /v1/pets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a pet
      operationId: DeletePet
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.DeletePetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Delete a pet
      tags:
      - pets v1
    get:
      consumes:
      - application/json
      description: Get a pet
      operationId: GetPet
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetPetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get a pet
      tags:
      - pets v1
    patch:
      consumes:
      - application/json
      description: Update a pet
      operationId: UpdatePet
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdatePetBody'
      - description: Owner ID
        in: body
        name: owner_id
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdatePetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Update a pet
      tags:
      - pets v1
schemes:
- http
swagger: "2.0"